+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Custom double tap to zoom for mobile
+
 ## [0.18.4] - 2025-03-15
 
 ### Added